"GridControlEmbedValue" should implement its own "IsValid" property
The class inherits from the "GridControlHtmlValue" class, which already has an implementation of the "IsValid" property. This implementation will however strip any HTML tags, and then check whether there is any text left. This isn't ideal for the "GridControlEmbedValue" class since the element to be embedded then will be stripped.