Added generic support for GridControl and GridEditor
Backported from the V9 version of this package.

For editor that has a configuration, the returned type is now "GridEditor<TConfig>" instead of just "GridEditor", which results in the "Config" property being of type "TConfig" of "IGridEditorConfig". "GridEditor<TConfig>" extends the "GridEditor" class, and "TConfig" must implement the "IGridEditorConfig" to ensure this doesnt' cause any breaking changes.

In the same way, controls will now be returned as "GridControl<TValue>" if is has a value - or even "GridControl<TValue, TConfig>" if the editor of the control also has a configuration. "GridControl<TValue, TConfig>" extends "GridControl<TValue>" whic extends "GridControl", and "TValue" must implement the "IGridControlValue" interface.